Relocate scaffold diff + framework files needed cluster (Phase 3, batch 4)
Issue #683 (parent #680, Phase 3) batch 4: migrate the scaffold-diff
cluster and the deterministic framework-files-needed cluster from
`turn.rs` to the sibling `scaffold_pipeline.rs` module.

Migrated to `scaffold_pipeline.rs` (`pub(super)`, 11 items):

* `ScaffoldDiffStatus` enum (NotScaffold / Changed / UnchangedOrMissing)
* `scaffold_file_snapshot`
* `scaffold_role_for_path`
* `scaffold_role_matches_artifact_role`
* `scaffold_diff_status`
* `scaffold_candidate_for_missing_role_from_snapshots`
* `scaffold_candidate_priority`
* `deterministic_framework_game_files_needed`
* `deterministic_framework_app_files_needed`
* `deterministic_framework_game_impl_path`
* `deterministic_support_target_relative`

Promoted in `turn.rs` to `pub(super)` so the moved helpers can call
back across the module boundary:

* `sha256_hex`
* `meaningful_workspace_files`
* `current_file_hash_for_relative_path`

`turn.rs` updates the `use super::scaffold_pipeline::{...}` blocks to
import the moved fns and types (production + cfg(test) test module).
Removed now-unused `ScaffoldArtifactRole` and
`implementation_quality_issue_for_request` imports from the file-level
imports; `implementation_quality_issue_for_request` is re-imported in
the test module via `super::super::quality`.

LOC delta:

* `turn.rs`: 31,656 → 31,455 (-201 LOC)
* `scaffold_pipeline.rs`: 263 → 473 (+210 LOC)
* Cumulative `turn.rs`: 39,489 → 31,455 (-8,034 LOC, -20.3%)

`pub(super)` limited / no facade re-export (DR3-001).
`turn.rs` remains the only in-crate consumer.

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
